# kyros-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kyros Backend",
    "description": "Wellness (Phase 1) + Clinic (Phase 2)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/wellness/logs/": {
      "post": {
        "tags": [
          "Wellness"
        ],
        "summary": "Create Log Entry",
        "operationId": "create_log_entry_v1_wellness_logs__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ]
      },
      "get": {
        "tags": [
          "Wellness"
        ],
        "summary": "List Log Entries",
        "operationId": "list_log_entries_v1_wellness_logs__get",
        "parameters": [
          {
            "name": "tracked_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tracked Item Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntryRead"
                  },
                  "title": "Response List Log Entries V1 Wellness Logs  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/wellness/logs/{log_id}": {
      "delete": {
        "tags": [
          "Wellness"
        ],
        "summary": "Delete Log Entry",
        "description": "Hard delete a log entry (user-corrected record, not adherence history).",
        "operationId": "delete_log_entry_v1_wellness_logs__log_id__delete",
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Id"
            }
          },
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "description": "16\u201364 char alphanumeric+dash device identifier for guest auth.",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9\\-]{16,64}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogEntryCreate": {
        "properties": {
          "tracked_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tracked Item Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "taken",
              "skipped",
              "snoozed",
              "logged_value",
              "acknowledged"
            ],
            "title": "Action"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "reminder_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Id"
          },
          "fire_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fire Key"
          },
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "snooze_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snooze Minutes"
          }
        },
        "type": "object",
        "required": [
          "tracked_item_id",
          "action",
          "occurred_at"
        ],
        "title": "LogEntryCreate"
      },
      "LogEntryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "tracked_item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tracked Item Id"
          },
          "reminder_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Id"
          },
          "fire_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fire Key"
          },
          "action": {
            "type": "string",
            "enum": [
              "taken",
              "skipped",
              "snoozed",
              "logged_value",
              "acknowledged"
            ],
            "title": "Action"
          },
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "tracked_item_id",
          "reminder_id",
          "fire_key",
          "action",
          "value",
          "note",
          "occurred_at",
          "source",
          "created_at"
        ],
        "title": "LogEntryRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
...
}